Several fixes, diagnostic plots and an example steering script for SVDDigitizer.
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@2933 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ svd/simulation/src/SVDSignal.cc
@@ -24,8 +24,8 @@ SVDSignal::operator()(double time)
   BOOST_FOREACH(SVDSignal::Wave wave, m_functions) {
     double wave_value = waveform(time, wave);
     wave_sum += wave_value;
-    m_particles[wave.m_particle] += wave_value;
-    m_truehits[wave.m_truehit] += wave_value;
+    m_particles[wave.m_particle] += static_cast<float>(wave_value);
+    m_truehits[wave.m_truehit] += static_cast<float>(wave_value);
   }
   return boost::make_tuple(wave_sum, boost::cref(m_particles), boost::cref(m_truehits));
 }
